refactoring: prepare for one call api
